LayoutTests/imported/w3c/web-platform-tests/editing/other/delete-without-unwrapping-first-line-of-child-block_method=BackspaceKey&lineBreak=preformat-expected.txt


ef
ghi

PASS Backspace at abc\\n[\\n<div id="child">]def<br>ghi</div>
PASS Typing "X" at abc\n[\n<div id="child">]def<br>ghi</div>
PASS Backspace at \\n[abc\\n<div id="child">]def<br>ghi</div>
PASS Typing "X" at \n[abc\n<div id="child">]def<br>ghi</div>
PASS Backspace at \\n[\\n<div id="child">]def<br>ghi</div>
PASS Typing "X" at \n[\n<div id="child">]def<br>ghi</div>
PASS Backspace at [\\n\\n<div id="child">]def<br>ghi</div>
PASS Typing "X" at [\n\n<div id="child">]def<br>ghi</div>
PASS Backspace at [abc\\n\\n<div id="child">]def<br>ghi</div>
PASS Typing "X" at [abc\n\n<div id="child">]def<br>ghi</div>
PASS Backspace at abc\\n[\\n<div id="child">d]ef<br>ghi</div>
PASS Typing "X" at abc\n[\n<div id="child">d]ef<br>ghi</div>
PASS Backspace at \\n[abc\\n<div id="child">d]ef<br>ghi</div>
PASS Typing "X" at \n[abc\n<div id="child">d]ef<br>ghi</div>
PASS Backspace at \\n[\\n<div id="child">d]ef<br>ghi</div>
PASS Typing "X" at \n[\n<div id="child">d]ef<br>ghi</div>
PASS Backspace at [\\n\\n<div id="child">d]ef<br>ghi</div>
PASS Typing "X" at [\n\n<div id="child">d]ef<br>ghi</div>
PASS Backspace at [abc\\n\\n<div id="child">d]ef<br>ghi</div>
PASS Typing "X" at [abc\n\n<div id="child">d]ef<br>ghi</div>
FAIL Backspace at abc\\n\\n<div id="child">[]def<br>ghi</div> assert_in_array: Backspace at abc\\n\\n<div id="child">[]def<br>ghi</div> should delete only the preceding empty line of the child <div> value "abc\n\n<div id=\"child\">ef<br>ghi</div>" not in array ["abc\n<div id=\"child\">def<br>ghi</div>", "abc<div id=\"child\">def<br>ghi</div>"]
FAIL Backspace at \\n\\n<div id="child">[]def<br>ghi</div> assert_equals: Backspace at \\n\\n<div id="child">[]def<br>ghi</div> should delete only the preceding empty line of the child <div> expected "\n<div id=\"child\">def<br>ghi</div>" but got "\n\n<div id=\"child\">ef<br>ghi</div>"
FAIL Backspace at \\n<div id="child">[]def<br>ghi</div> assert_equals: Backspace at \\n<div id="child">[]def<br>ghi</div> should delete only the preceding empty line of the child <div> expected "<div id=\"child\">def<br>ghi</div>" but got "\n<div id=\"child\">ef<br>ghi</div>"
FAIL Backspace at <b>abc\\n\\n</b></b><div id="child">[]def<br>ghi</div> assert_in_array: Backspace at <b>abc\\n\\n</b></b><div id="child">[]def<br>ghi</div> should delete only the preceding empty line of the child <div> (<b> should stay) value "<b>abc\n\n</b><div id=\"child\">ef<br>ghi</div>" not in array ["<b>abc\n</b><div id=\"child\">def<br>ghi</div>", "<b>abc</b><div id=\"child\">def<br>ghi</div>"]
FAIL Backspace at <b>\\n</b><div id="child">[]def<br>ghi</div> assert_equals: Backspace at <b>\\n</b><div id="child">[]def<br>ghi</div> should delete only the preceding empty line (including the <b>) of the child <div> expected "<div id=\"child\">def<br>ghi</div>" but got "<b>\n</b><div id=\"child\">ef<br>ghi</div>"
FAIL getTargetRanges() for Backspace at abc\\n[\\n<div id="child">]def<br>ghi</div> should return a range before the child <div> assert_in_array: value "[{(#text \"abc\\\\n\", 4) - (<div id=\"child\">, 0)}]" not in array ["[{(#text \"abc\\\\n\", 4) - (<div contenteditable=\"\">, 1)}]", "[{(#text \"abc\\\\n\", 3) - (<div contenteditable=\"\">, 1)}]", "[{(#text \"abc\\\\n\", 4) - (#text \"abc\\\\n\", 5)}]", "[{(#text \"abc\\\\n\", 3) - (#text \"abc\\\\n\", 5)}]"]
FAIL getTargetRanges() for Typing "X" at abc\n[\n<div id="child">]def<br>ghi</div> should return a range ending in the child <div> assert_in_array: value "[{(#text \"abc\\\\n\", 4) - (<div id=\"child\">, 0)}]" not in array ["[{(#text \"abc\\\\n\", 4) - (#text \"Xdef\", 0)}]", "[{(#text \"abc\\\\n\", 3) - (#text \"Xdef\", 0)}]"]
FAIL getTargetRanges() for Backspace at \\n[abc\\n<div id="child">]def<br>ghi</div> should return a range before the child <div> assert_in_array: value "[{(#text \"\\\\n\", 1) - (<div id=\"child\">, 0)}]" not in array ["[{(#text \"\\\\n\", 1) - (<div contenteditable=\"\">, 1)}]", "[{(#text \"\\\\n\", 1) - (#text \"\\\\n\", 5)}]"]
FAIL getTargetRanges() for Typing "X" at \n[abc\n<div id="child">]def<br>ghi</div> should return a range ending in the child <div> assert_in_array: value "[{(#text \"\\\\n\", 1) - (<div id=\"child\">, 0)}]" not in array ["[{(#text \"\\\\n\", 1) - (#text \"Xdef\", 0)}]"]
FAIL getTargetRanges() for Backspace at \\n[\\n<div id="child">]def<br>ghi</div> should return a range before the child <div> assert_in_array: value "[{(#text \"\\\\n\", 1) - (<div id=\"child\">, 0)}]" not in array ["[{(#text \"\\\\n\", 1) - (<div contenteditable=\"\">, 1)}]", "[{(#text \"\\\\n\", 1) - (#text \"\\\\n\", 2)}]"]
FAIL getTargetRanges() for Typing "X" at \n[\n<div id="child">]def<br>ghi</div> should return a range ending in the child <div> assert_in_array: value "[{(#text \"\\\\n\", 1) - (<div id=\"child\">, 0)}]" not in array ["[{(#text \"\\\\n\", 1) - (#text \"Xdef\", 0)}]"]
FAIL getTargetRanges() for Backspace at [\\n\\n<div id="child">]def<br>ghi</div> should return a range before the child <div> assert_in_array: value "[{(#text \"\\\\n\\\\n\", 0) - (<div id=\"child\">, 0)}]" not in array ["[{(#text \"\\\\n\\\\n\", 0) - (<div contenteditable=\"\">, 1)}]", "[{(#text \"\\\\n\\\\n\", 0) - (#text \"\\\\n\\\\n\", 2)}]"]
FAIL getTargetRanges() for Typing "X" at [\n\n<div id="child">]def<br>ghi</div> should return a range ending in the child <div> assert_in_array: value "[{(#text \"\\\\n\\\\n\", 0) - (<div id=\"child\">, 0)}]" not in array ["[{(#text \"\\\\n\\\\n\", 0) - (#text \"Xdef\", 0)}]"]
FAIL getTargetRanges() for Backspace at [abc\\n\\n<div id="child">]def<br>ghi</div> should return a range before the child <div> assert_in_array: value "[{(#text \"abc\\\\n\\\\n\", 0) - (<div id=\"child\">, 0)}]" not in array ["[{(<div contenteditable=\"\">, 0) - (<div contenteditable=\"\">, 1)}]", "[{(#text \"abc\\\\n\\\\n\", 0) - (<div contenteditable=\"\">, 1)}]", "[{(#text \"abc\\\\n\\\\n\", 0) - (#text \"abc\\\\n\\\\n\", 5)}]"]
FAIL getTargetRanges() for Typing "X" at [abc\n\n<div id="child">]def<br>ghi</div> should return a range ending in the child <div> assert_in_array: value "[{(#text \"abc\\\\n\\\\n\", 0) - (<div id=\"child\">, 0)}]" not in array ["[{(#text \"abc\\\\n\\\\n\", 0) - (#text \"Xdef\", 0)}]"]
PASS getTargetRanges() for Backspace at abc\\n[\\n<div id="child">d]ef<br>ghi</div> should return a range ends at start of the child <div>
PASS getTargetRanges() for Typing "X" at abc\n[\n<div id="child">d]ef<br>ghi</div> should return a range ends at start of the child <div>
PASS getTargetRanges() for Backspace at \\n[abc\\n<div id="child">d]ef<br>ghi</div> should return a range ends at start of the child <div>
PASS getTargetRanges() for Typing "X" at \n[abc\n<div id="child">d]ef<br>ghi</div> should return a range ends at start of the child <div>
PASS getTargetRanges() for Backspace at \\n[\\n<div id="child">d]ef<br>ghi</div> should return a range ends at start of the child <div>
PASS getTargetRanges() for Typing "X" at \n[\n<div id="child">d]ef<br>ghi</div> should return a range ends at start of the child <div>
PASS getTargetRanges() for Backspace at [\\n\\n<div id="child">d]ef<br>ghi</div> should return a range ends at start of the child <div>
PASS getTargetRanges() for Typing "X" at [\n\n<div id="child">d]ef<br>ghi</div> should return a range ends at start of the child <div>
PASS getTargetRanges() for Backspace at [abc\\n\\n<div id="child">d]ef<br>ghi</div> should return a range ends at start of the child <div>
PASS getTargetRanges() for Typing "X" at [abc\n\n<div id="child">d]ef<br>ghi</div> should return a range ends at start of the child <div>
